Guess host OS based on uname for non-Debian systems. Closes: #637363
Deboostrap only cares if the host OS is Linux, Hurd, or kFreeBSD.
The actual architecture used on the host doesn't matter, as long as
the target arch can run in the chroot.

This does not address running debootstrap on a FreeBSD system in order
to build a kFreeBSD chroot. That seems to have already been not working,
since debootstrap checked for kfreebsd, and not freebsd.