point users at the log file on error, try to show a package name too
When installation or configuration of a package fails, output a message
that points the user to the log file[1]. Attempt to grep out the first
package that dpkg failed on and show its name too. Closes: #472704

Sample:

W: Failure while configuring base packages.  This will be re-attempted up to five times.
W: See sid/debootstrap.log for details (possibly the package libept1 is at fault)

This relies on the LANG=C currently set for the debootstrap run.
Changes to the "dpkg: error processing $PACKAGE" message would break
the package name extractor, but it'd fall back to not showing a
package name.

This adds cut to the set of utilities debootstrap uses, which shouldn't
be a problem.

[1] In d-i, debootstrap doesn't know where the log file is, but then
this message goes to the same syslog file, so I didn't add any special
handling for this case.